requirements.txt
django>=4.2.20
djangorestframework==3.15.2
django-extensions==3.2.3
psycopg==3.2.5
pydotplus==2.0.2
Pillow==10.4.0
configparser==7.1.0
djangorestframework-simplejwt==5.3.1
django-allauth==65.7.0
channels==4.2.2
channels_redis==4.2.1